+++ src/layer.h
@@ -58,8 +58,8 @@ Layer* createLayer(LAYER_TYPE type, int size, void (*activation)(Matrix*)){
     layer->type = type;
     layer->size = size;
     layer->activation = activation;
-    double** row = (double**)malloc(sizeof(double*));
-    row[0] = (double*)malloc(sizeof(double) * size);
+    float** row = (float**)malloc(sizeof(float*));
+    row[0] = (float*)malloc(sizeof(float) * size);
     layer->input = createMatrix(1, size, row);
     return layer;
 }
@@ -68,14 +68,14 @@ Connection* createConnection(Layer* from, Layer* to){
     Connection* connection = (Connection*)malloc(sizeof(Connection));
     connection->from = from;
     connection->to = to;
-    double** weights_data = (double**)malloc(sizeof(double*) * from->size);
+    float** weights_data = (float**)malloc(sizeof(float*) * from->size);
     int i;
     for (i = 0; i < from->size; i++){
-        weights_data[i] = (double*)malloc(sizeof(double) * to->size);
+        weights_data[i] = (float*)malloc(sizeof(float) * to->size);
     }
     connection->weights = createMatrix(from->size, to->size, weights_data);
-    double** bias_data = (double**)malloc(sizeof(double*));
-    bias_data[0] = (double*)malloc(sizeof(double) * to->size);
+    float** bias_data = (float**)malloc(sizeof(float*));
+    bias_data[0] = (float*)malloc(sizeof(float) * to->size);
     connection->bias = createMatrix(1, to->size, bias_data);
     return connection;
 }
